+++ contracts/interfaces/YearnGovernanceInterface.sol
@@ -3,15 +3,15 @@
 pragma solidity 0.6.12;
 
 interface YearnGovernanceInterface {
-    function stake(uint256 amount) external;
+  function stake(uint256 amount) external;
 
-    function withdraw(uint256 amount) external;
+  function withdraw(uint256 amount) external;
 
-    function voteFor(uint id) external;
+  function voteFor(uint256 id) external;
 
-    function voteAgainst(uint id) external;
+  function voteAgainst(uint256 id) external;
 
-    function balanceOf(address) external view returns (uint);
+  function balanceOf(address) external view returns (uint256);
 
-    function voteLock(address) external view returns (uint);
+  function voteLock(address) external view returns (uint256);
 }
